+++ CHANGELOG.md
@@ -4,5 +4,8 @@ All notable changes to this project will be documented in this file.
 This format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
+- Save the size and position of the window
+- Check if the URL is a valid YouTube Channel URL
+
 ## [1.0.0] - 2025-03-23
 - First version of the application with all the essential features.